temp_requirements.txt
absl-py==2.1.0
acres==0.2.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
astunparse==1.6.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.7.0
boto3==1.35.54
botocore==1.35.54
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
configobj==5.0.9
configparser==6.0.1
contourpy==1.3.1
cryptography==42.0.5
curlify==2.2.1
customerio==2.1
cycler==0.12.1
distro==1.9.0
editor==1.6.6
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.12.0
gast==0.6.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.68.1
gspread==6.1.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.12.1
htmlmin==0.1.12
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.6
ImageHash==4.3.1
importlib_resources==6.4.5
inquirer==3.2.4
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.2.1
jiter==0.8.2
jmespath==1.0.1
joblib==1.1.1
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
looseversion==1.3.0
lxml==5.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.10.0
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.4.1
multidict==6.0.5
multimethod==2.0
namex==0.0.8
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.11.0.86
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
phik==0.12.4
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
prov==2.0.1
psycopg2==2.9.9
puremagic==1.28
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.4
PyDrive2==1.21.1
Pygments==2.18.0
PyMuPDF==1.25.1
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.8.0
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
readchar==4.0.6
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
runs==1.2.2
s3transfer==0.10.3
safetensors==0.5.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.18.1
setuptools==69.5.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
StrEnum==0.4.15
sweetviz==2.3.1
tangled-up-in-unicode==0.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.21.0
tqdm==4.67.1
traits==6.4.3
transformers==4.47.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
visions==0.7.4
waitress==3.0.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.9
xmod==1.8.1
yarl==1.9.4
